Disable deprecation notices for PGO by default
PGO optimization is meant for production, so we should use the
recommended `error_reporting` setting for such environments.  This is
especially relevant since PHP is deprecating lots of stuff, and we do
not want to optimize any code paths which are triggered by deprecation
notices.

Of course, individual apps may override that, but choosing a reasonable
default is a first step.